# docs/static/openapi.yaml
openapi: 3.0.1
info:
  title: Szampchat server api
  version: "0.1"
servers:
- url: http://localhost:8081/api
  description: Local server
paths:
...
  /communities:
    get:
      tags:
      - Community
      summary: Get user's communities
      description: Retrieves communities of logged in user
      operationId: getUserCommunities
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Community"
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
    post:
      tags:
      - Community
      summary: Create community
      description: |-
        Creates new community with current user as a owner.
        New members can be invited using (link to other endpoints)
      operationId: createCommunity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityCreateDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Community"
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
...
  /communities/{communityId}/join:
    post:
      tags:
      - Community
      summary: Join community
      description: |-
        Uses invitation link (it's id) to join community as current user.
        This action result in expiration of invitation. (?)
      operationId: joinCommunity
      parameters:
      - name: communityId
        in: path
        description: ID of community
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRequestDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityMember"
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
        "403":
          description: User is already a member of community or resource is not available
      security:
      - OAuthSecurity: []
  /communities/{communityId}/invite:
    post:
      tags:
      - Community
      summary: Create invite link
      description: |-
        Creates temporary or permanent(?) invitation link to community.
        This invitation can then be shared with any user, who can then join given community
      operationId: inviteToCommunity
      parameters:
      - name: communityId
        in: path
        description: ID of community
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponseDTO"
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
        "403":
          description: User is not an owner of community or resource is not available
      security:
      - OAuthSecurity: []
...
  /channels/{channelId}/messages:
    get:
      tags:
      - Message
      summary: Get channel's messages
      operationId: getMessages
      parameters:
      - name: channelId
        in: path
        description: ID of channel
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      - name: before
        in: query
        description: Get messages before this message Snowflake ID
        required: false
        schema:
          type: integer
          format: int64
        example: 21388369783160832
      - name: limit
        in: query
        description: Max number of messages
        required: false
        schema:
          type: integer
          format: int32
        example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageDTO"
        "419":
          description: User not created
        "403":
          description: User is not a member of channel's community or resource is
            not available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
    post:
      tags:
      - Message
      summary: Create message
      operationId: createMessage
      parameters:
      - name: channelId
        in: path
        description: ID of channel
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageCreateDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "419":
          description: User not created
        "403":
          description: User is not a member of channel's community or resource is
            not available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
...
  /communities/{communityId}:
    get:
      tags:
      - Community
      summary: Get community
      operationId: getCommunity
      parameters:
      - name: communityId
        in: path
        description: ID of community
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityDTO"
        "419":
          description: User not created
        "403":
          description: User is not a member of community or resource is not available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
    delete:
      tags:
      - Community
      summary: Delete community
      description: Removes community
      operationId: deleteCommunity
      parameters:
      - name: communityId
        in: path
        description: ID of community
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      responses:
        "204":
          description: No Content
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
        "403":
          description: User is not an owner of community or resource is not available
      security:
      - OAuthSecurity: []
    patch:
      tags:
      - Community
      summary: Edit community
      description: Edits community
      operationId: editCommunity
      parameters:
      - name: communityId
        in: path
        description: ID of community
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Community"
        required: true
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Community"
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
        "403":
          description: User is not an owner of community or resource is not available
      security:
      - OAuthSecurity: []
...
  /channels/{channelId}/messages/{messageId}:
    delete:
      tags:
      - Message
      summary: Delete message
      operationId: deleteMessage
      parameters:
      - name: channelId
        in: path
        description: ID of channel
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      - name: messageId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "419":
          description: User not created
        "403":
          description: User is not a member of channel's community or resource is
            not available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
    patch:
      tags:
      - Message
      summary: Edit message
      operationId: editMessage
      parameters:
      - name: channelId
        in: path
        description: ID of channel
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      - name: messageId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditMessageDTO"
        required: true
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "419":
          description: User not created
        "403":
          description: User is not a member of channel's community or resource is
            not available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
...
  /communities/{communityId}/info:
    get:
      tags:
      - Community
      summary: Get full community
      description: |-
        Retrieve all required information about community.
        Ideally it should be used once when opening community
        and then updated using websocket events
      operationId: getFullCommunityInfo
      parameters:
      - name: communityId
        in: path
        description: ID of community
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullCommunityInfoDTO"
        "419":
          description: User not created
        "403":
          description: User is not a member of community or resource is not available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
components:
  schemas:
    Channel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        communityId:
          type: integer
          format: int64
        type:
          type: string
          enum:
          - TEXT_CHANNEL
          - VOICE_CHANNEL
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID of user
          format: int64
          example: 4501678724218880
        name:
          type: string
          description: User name
          example: User123
        imageUrl:
          type: string
          description: Relative path to image (or uuid)
          example: img/userimage.jpg
        description:
          type: string
          description: Short message which describes user
          example: My description
    CommunityCreateDTO:
      type: object
      properties:
        name:
          type: string
          description: Name of community
          example: My community
    Community:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        ownerId:
          type: integer
          format: int64
        imageUrl:
          type: string
        basePermissions:
          $ref: "#/components/schemas/Permissions"
    Permissions:
      type: object
      properties:
        permissionData:
          type: integer
          format: int32
...
    PermissionOverwrites:
      type: object
      properties:
        permissionOverwriteData:
          type: integer
          format: int64
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        permission:
          $ref: "#/components/schemas/PermissionOverwrites"
        community:
          type: integer
          format: int64
    JoinRequestDTO:
      type: object
      properties:
        invitationId:
          type: integer
          description: Valid ID of invitation
          format: int64
          example: 29754971884879870
    CommunityMember:
      type: object
      properties:
        communityId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
    InvitationResponseDTO:
      type: object
      properties:
        link:
          type: string
          description: Invitation link to community
          example: community/4501678724218880/join/29754971884879870
...
    MessageCreateDTO:
      type: object
      properties:
        text:
          type: string
        communityId:
          type: integer
          format: int64
        respondsToMessage:
          type: integer
          format: int64
        gifLink:
          type: string
    Message:
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        updated_at:
          type: string
          format: date-time
        respondsToMessage:
          type: integer
          format: int64
        gifLink:
          type: string
        channelId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
...
    EditMessageDTO:
      type: object
      properties:
        text:
          type: string
    CommunityDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID of community
          format: int64
          example: 4501678724218880
        name:
          type: string
          description: Name of community
          example: My community
        ownerId:
          type: integer
          description: ID of community's owner
          format: int64
          example: 29754971884879872
        imageUrl:
          type: string
          example: img/myimage.png
        basePermissions:
          $ref: "#/components/schemas/Permissions"
    CommunityMemberRolesDTO:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserDTO"
        roles:
          uniqueItems: true
          type: array
          description: Roles of user in given community
          items:
            type: integer
            description: ID of role
            format: int64
            example: 8518649245597696
    FullCommunityInfoDTO:
      type: object
      properties:
        community:
          $ref: "#/components/schemas/CommunityDTO"
        channels:
          type: array
          items:
            $ref: "#/components/schemas/Channel"
        members:
          type: array
          items:
            $ref: "#/components/schemas/CommunityMemberRolesDTO"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    MessageAttachmentDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID of attachment
          format: int64
          example: 4501678724218880
        size:
          type: integer
          description: Size of file in bytes
          format: int32
          example: 2048
        name:
          type: string
          description: ID of attachment
          example: myfile.txt
    MessageDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID of message
          format: int64
          example: 4501678724218880
        text:
          type: string
          description: Content of message
          example: My message content
        updated_at:
          type: string
          description: When message was updated
          format: date-time
        respondsToMessage:
          type: integer
          description: ID of message which this message responds to
          format: int64
          example: 4501678724218880
        gifLink:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/MessageAttachmentDTO"
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionPreviewDTO"
        channelId:
          type: integer
          description: ID of message's text channel
          format: int64
          example: 4501678724218880
        userId:
          type: integer
          description: ID of user that sent message
          format: int64
          example: 4501678724218880
    ReactionPreviewDTO:
      type: object
      properties:
        emoji:
          type: string
          description: Emoji in unicode
          example: 👍
        count:
          type: integer
          description: How many users used this reaction
          format: int32
          example: 10
        me:
          type: boolean
          description: If current user reacted to the message
          example: true
  securitySchemes:
    OAuthSecurity:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://url.com/auth
          scopes: {}