# server/src/main/java/com/szampchat/server/role/dto/RoleDTO.java
package com.szampchat.server.role.dto;

import com.szampchat.server.permission.data.PermissionOverwrites;

public class RoleDTO {
    private Long id;
    private String name;
    private PermissionOverwrites permissionOverwrites;
    private Long community;
}